+++ CHANGELOG.md
@@ -278,3 +278,4 @@
 * Updated CREDITS
 * Options panel now displays game name and versions of game, core, and engine
 * Options subpanels can be added with a builder function instead of prefab (and always use a build func internally)
+* RpgChangeWeapon message is now RpgEquipmentChanged, more accurately reflecting its actual function